=== BEGIN compat-modules-2 === Move createPortal to its own module (-5 B)
